refactor: unify episode card system into EpisodeVerticalCard and EpisodeCompactCard
Replace 5 duplicate card implementations (EpisodeRailCard, LibraryEpisodeCard,
PodcastEpisodeCard, QueueItemCard, and their inline variants) with two shared
reusable components in CardComponents.swift:

- EpisodeVerticalCard: 200pt-wide card for horizontal scroll rails with artwork
  on top, explanation tag, metadata, play/queue circle buttons
- EpisodeCompactCard: full-width horizontal card for vertical lists with 56x56
  artwork, podcast title, metadata, play circle, optional rank and remove button

Consistent 36pt icon circles, standardized spacing, and unified metadata format
across Home, Library, and Queue views. Net removal of ~355 lines of duplicated
card code.
